refactor: change package name/location
