+++ src/main/java/com/neoga/boltauction/security/util/CustomAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.neoga.boltauction.security.config;
+package com.neoga.boltauction.security.util;
 
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
