+++ src/main/java/com/neoga/boltauction/util/S3Uploader.java
@@ -1,4 +1,4 @@
-package com.neoga.boltauction.s3;
+package com.neoga.boltauction.util;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
